Improved cursor by centering at origin and then translating
Cursors appear to be defined with a points list that should stay in the
positive quadrant. This is somewhat unnatural for defining a symmetric
cursor (such as our 4 arrow move cursor). This change recodes the cursor
to be symmetric about the origin and then translates it to be in the
positive quadrant. This change also implemented a thinner cursor which
is more easily changed with a few integer parameters:

   cw2: Cursor width / 2
   ch2: Cursor height / 2
   clw: Arrow line width (actually half of the width)
   caw: Arrow width (actually half of arrow head width)
   cal: Arrow head length

These can be adjusted to get various cursor arrow shapes.