Updated the modes to better handle the center drawing mode.
The Center Drawing mode draws from the center cursor as the image is
moved around below it. It's somewhat like using a sewing machine or a
table-based tool (like a jig saw). As the image is pushed upward (with
the mouse) the line is drawn downward with respect to the image itself.

Note that the direction of movement could be reversed so that moving up
will cause the image to move down, but the direction of drawing would be
"up" with respect to the image. This isn't implemented in this version
because the moving itself is handled in a lower layer (ZoomPanLib.java),
and that *might* need to be modified to implement the reversal.