Added basic XML parsing of a series file. No data extracted yet.
Note that because we don't have a DTD for the series files, but they are
produced (by the original Reconstruct) with a "<!DOCTYPE", the DOCTYPE
must be removed prior to passing the file to Java's XML parsing code.
Otherwise, Java's XML parsing code tries to check the XML against the
DTD and fails.

The code at the top of parse_series_xml_file() reads the XML file line
by line and removes any lines that match "<!DOCTYPE" at the beginning of
the line. The lines are then converted to a StringBufferInputStream to
be passed as a stream to the parser (dBuilder.parse).