Added code to search for section files starting with the series file name.
The current Reconstruct doesn't keep a master list of section files
inside the series file itself. Instead, all files in the directory of
the series file with the same name (excluding any suffix) and ending in
a number is considered a section file.

This code currently doesn't exclude files with non-numeric extensions
except for the ".ser" file itself.