Added more proper filtering for section files.
A series file name will be of the form:

   any_name.ser

All of the section files will then be of the form:

   any_name.1
   any_name.2
    :
   any_name.9
   any_name.10
    :

The previous version just matched against any file that started with
"any_name" but explicitly removed "any_name.ser". This code is a bit
more explicit and searches for "any_name" that ends with a series of
numerals (0-9) after the final ".". This might also include other file
names with multiple "dots" (any_name.bad_file.123), but that's not
expected at this point in the development. A slightly better regx will
eventually fix this as well.