Added right click to switch drawing modes. Preparing for series scroll.
The original Java demonstration used the scroll wheel for zooming, but
the original reconstruct used the scroll wheel for moving through the
series. This version uses the right mouse button to toggle between the
draw and move modes and only uses the scrollwheel for zooming when in
draw mode. When in move mode, the scrollwheel will change between the
sections in the series.