Isolated the dump_doc functionality to be used for testing.
The earlier versions dumped the XML document while reading. This version
adds an additional dump_doc function so it can be done then and also on
command.

It's not clear yet if it would be best to work with the XML data as it
is inside the Document class or if it should be tranlated into another
set of more specific classes. This may depend on how easy the the
Document class is to use.