Replaced object.<KEY_CODE> with class.<KEY_CODE> to suppress warnings.
The newer versions of Java seem to prefer referencing static members of
a class through the class rather than through an object of the class.

This can be clearer because it shows that the member is static, but it
might also be more error prone since future changes that change the class
type of the object to a different class will not automatically call or
catch the change to the static member.

As an example consider the following comparison:

  e.getKeyCode() == e.VK_PAGE_DOWN

In this case, "e" happens to be an object of class "KeyEvent" which
contains the static member VK_PAGE_DOWN. If the class type of "e" were
changed to represent a different object then the following might happen:

  1. The new class might also have a VK_PAGE_DOWN member with a
     different value than in KeyEvent. The code would still work
     using this new - and more likely appropriate - value rather
     than the one from its old class type of KeyEvent.

  2. The new class might NOT have a VK_PAGE_DOWN member which would
     generate a compile time error indicating a potential logic error.

Either of those are likely to lead to properly working code. But using
the class reference to VK_PAGE_DOWN (rather than the object reference)
would not flag an error if the type of the object were changed to either
a class without a VK_PAGE_DOWN member or to a class with VK_PAGE_DOWN
defined to be a different value. Both of those cases might cause subtle
and difficult to detect errors.

Having said that, in order to suppress the warnings, the changes were
made to reference the static values by class rather than by object.