First attempts to pull specific data from XML document (index).
This version prints the current section index (stored in the Series) and
each individual section index (stored in each Section). The syntax in
both cases is the same:

  ...doc.getDocumentElement().getAttributes().getNamedItem("index")

These kinds of calls might be candidates for convenience functions in
the Series and/or Section classes. Maybe they should inherit from a
common base class with that kind of functionality.