Added the jshell_setup file for debugging via jshell.
Java 9 has a new program named "jshell" that provides an interactive
environment somewhat similar to the Python command shell. The jshell
command can take a "script" file that initializes the environment by
running a number of jshell commands.

The jshell_setup file is intended to set up the jshell environment for
working with Reconstruct. It currently opens (effectively "imports") a
few java files, prints the loaded types, opens a series file, creates a
new SeriesClass (which also creates the Section classes). This is very
helpful for working interactively with the XML document model.