Commented out printing and decoupled scroll_wheel_position in draw.
This version will open a normal Reconstruct series file (XML) and load
all of the images. The images can be scrolled through and zoomed and
panned. The right mouse button toggles between drawing (where the scroll
wheel controls depth in the stack) and dragging (where the scroll wheel
controls the zoom). Beware, however, that this version loads all the
images at once and this may take a very very long time if you have many
large images in your series. It's best to test on either small images
or very few sections.

A future improvement will use threads to load images in the background
and have some reasonable strategy for anticipating which ones to load.

Also note that the traces are all still in one layer, and that this
version does not save any of the traces to file. In fact, the XML files
are not modified at all by this version.