BezierTracing: Allow Bezier handles to have different lengths.
A Bezier segment is defined by two end points and two control points
(also known as "handles").

A Bezier curve (in the context of this program) is an ordered list of
Bezier segments such that neighboring segments in that list share end
points, each end point is only shared with one other segment, and all
of the segments are joined either directly or indirectly to all others.

All of the segments in a curve form a continuous trace. However, there
is no guarantee that they are joined smoothly (without sharp angles or
discontinuities in their slopes). This may be desired in some cases
(where a sharp corner is needed) or undesired in others (where the
surface is intended to always be smooth).

This commit enforces a common slope between handles on the same point,
but it does allow different lengths on each side of the shared point.
This allows one side to be very nearly zero resulting in a sharp point.

Eventually, the greatest flexibility will come from providing all of:

  1. Shared Slope / Independent Length
  2. Shared Slope / Shared Length
  3. Independent Slope / Independent Length

The current design supports (1) above.