Java: Modified SectionClass to detect both images and traces.
The XML for a single section file contains a series of Transforms.

It appears that each Transform contains either:

   One Image element  and  One Contour element

 - or -

   One or more Contour elements (without any Images)

The "Contour" contained in the first case appears to be the perimeter of
the image. The Contours in the second case are the traces themselves.

This commit searches each Transform for the presence of an "Image"
element, and uses that as a flag to determine what to pull out of the
Transform (an Image or a series of Contours as traces).

Note that the traces themselves are not pulled out yet.