Added code to use actual menu items rather than menu names.
The Java menu system passes an ActionEvent to the actionPerformed method
of the listening class. The ActionEvent contains a number of fields
including the menu item (source) of the event, and the action command
(typically the menu item text). Early versions of the Windowing toolkit
used the action command (typically a string) to determine which action
to take. This is commonly sufficient unless a menu tree might contain
multiple items with the same name (typically at the "leaves"). In this
case, it's better to save the menu items as they are created, and then
use the "source" field to compare with the created items. This has the
advantage of guaranteeing uniqueness even when menu items have the same
names. However, it does require storing all of the menu items in unique
variables to compare to the source of any event. While slightly more
work, the second approach guarantees that there are no name conflicts
because they will be caught at compile time.

This commit begins the transition from using string names to using saved
menu items for comparison. In this case, the "open_series_menu_item" has
been created to store the menu item (at creation time) and allow for
comparison (at event handling time). The "Open..." menu item was chosen
as the first implementation because there may be multiple "leaf" menu
items that "Open" something. This would allow them all to use the same
string name ("Open...") at different parts of the menu tree.