Converted remaining command strings to menu items.
Note that this seems to limit the ability to put the same menu items at
different points in the menu tree. Previously (using command strings),
the "Dump" and "Clear" menu items were in both the "Debug" menu and in
the "Mode" menu. However, placing them in the "Mode" menu seemed to
remove them from the "Debug" menu. For that reason, they were commented
out of the "Mode" menu in this revision.