Added more output to the main() for SectionClass.java
As mentioned in commit be0089bc49fe78f3cca232dcc491802baf1aeef9, there
is a problem using JShell with the current class structures. This is
extremely unfortunate because JShell would be ideal for exploring the
structures returned by the XML library (much of the existing XML code
was written based on that exploration). As an alternative, this commit
adds code in the "main()" function for the SeriesClass to print out some
of that XML structure data. The "main" function for the SeriesClass
would not normally be called anyway, and it is a good place for unit
testing and debugging.