Added Optional DTD handling in Java version and Updated DTDs
Previous versions had been non-validating (setValidating(false)).
They had also used a special read loop that filtered out the XML
lines containing the DTD reference.

This version attempts to read (parse) the XML directly with the
DTD validation enabled. However, if that fails, then it reverts
to the previous behavior of filtering the DTD reference from the
XML file.

Note that it's not clear how to best reference the DTD files from
within the .jar file, so the DTD files will generally not be found
unless java is launched with the working directory set to the location
of the .jar file itself (assuming that the DTDs are copied or linked
to that location). Eventually, it would be best to have the DTDs
contained in the .jar file so they can be found however the .jar
file is launched.

In order to validate current Reconstruct XML files (generated by the
"C" version of Reconstruct), the following fields needed to be added
to the DTD files:

Additions to the "Series" ELEMENT in "series.dtd":

   ContourMaskWidth     %SFFloat;   "0"
   smoothingLength      %SFFloat;   "7"

Additions to the "Image" ELEMENT in "section.dtd":

   red         %SFBool;    "true"
   green       %SFBool;    "true"
   blue        %SFBool;    "true"

With these changes, the current example XML files were validated.
These defaults were taken from one of the example XML files.