Added a check to catch section file names that end with ".0"
Previous examples all started with section number 1 so the code was
written for that case. However, some BBCHZ series seem to start with
section number 0. This caused an error in the sorting routine which
padded all numeric extensions to be the same width before sorting.

If, for example, there were a few hundred files, then file.1 would
become file.001, and file.15 would become file.015. Then the names
could be easily sorted. But to recover the original file names, the
leading zeros needed to be removed again. The algorithm used worked
great as long as ALL of the extension characters were 0 (zero). This
commit handles the case where all digits are zero and returns a file
name of "file.0" after sorting.