Added code to only report missing image files once.
The calibration "section" is currently treated like all other sections.
If the associated image is missing, it will be reported with a pop up.

Since the calibration image is sometimes missing from a data set, this
will cause the pop up message to be repeated with every view of this
first (#0) section. This commit keeps track of such missing files and
only generates the pop up once for each missing image.