Partially Functional Series Reader (Transforms not yet correct)
This version can read a Reconstruct Series file and load all of the
images and traces. However, the transforms are not handled properly,
and this causes the traces to be displayed in the wrong location.

This version is good for testing the responsiveness of reading multiple
images and scrolling through them in Java. Note that all of the XML
files are read and parsed initially which can take a few seconds.