Java: Added ability to partially create a new series from images.
This version of Reconstruct.java supports the "Series / New..." menu
option and the "Series / Import Images..." menu option.

The "Series/New" option prompts for the name of a series file and
creates the series file itself (with no sections).

The "Series/Import Images..." menu option creates all the section files
for the images selected. Note that the transforms are not of the correct
size for the images (currently defaults to 480x480). This size, however,
is not currently used by this code, so the results appear fine when used
with the Java version.

The menu options are not yet properly "grayed out" with this version. So
it is possible to try to import images without having opened (or created
a series file). Any attempt to do this is simply ignored.