Added (unused) Contour class and default scaling for BBCHZ series.
The current strokes (traces) are represented by an array of arrays of
points. This was a good start, but the arrays have no ability to hold
other attributes like trace names and trace colors. This commit
introduces a new class to represent a "Contour" which can hold a trace.
The name "Contour" was used to reflect the similar use in the C version.

This version also includes default settings for the BBCHZ transform
traces transform (not the same as the calibration transform). This
allows the BBCHZ series to be opened and to display properly without
adjusting the manual scaling values. This is a temporary work around
until the proper transforms can be implemented.