Added ability to show no traces and to purge images.
This commit added a "none" option to the lines menu to not draw traces
at all.

This commit also added the ability to purge images. This program
currently loads images as they're needed and keeps them in memory until
an "Out of Memory" error is detected. The logic for removing images in
that case is still a work in progress, so the "Purge Images" option was
added as a manual way to remove all images. They will again be loaded on
demand after the purge.