Java Version: Moved drawing into Contour class, Added filling for mode<0
The ContourClass now holds what it needs to do the drawing, so that
functionality has been moved into the ContourClass itself.

The Reconstruct Contours have a "mode" which appears to be negative when
filling is needed. This version fills whenever mode is less than 0. It's
not clear if this is the proper definition of that value.