Use Java's GeneralPath class for speed, and only fill mode = -13
Earlier versions drew each path segment by segment. This version uses
Java's ability to draw "GeneralPath" objects. This is likely much faster
than looping in Java and it's consistent with the "fill" code (which
also uses a GeneralPath).

This version only fills when mode = -13. The previous version assumed
that all negative mode values were filled, but that doesn't appear to be
the case. It's not yet clear what the "mode" values mean.