Added TransformClass with ability to handle offsets for traces.
It appears that trace offsets are a common form of transform that need
to be handled for the series being tested.

This commit adds the TransformClass and supports transforms with type
(dim) of 0 or 1. Higher transforms will be decoded, but the drawing code
does not use any higher order terms at this time.

A number of contours in the test series differ only by an offset. This
change displays them properly.

This commit also supports the "hidden" value in a contour.