Modified saving of Series file to be identical to original.
The original version of Reconstruct puts the index and viewport
attributes on the opening "Series" line. It also splits the color lines
by commas. This version added special code to handle those two special
cases.

The goal is to produce byte-for-byte identical XML files for easier
comparison and interoperability with the earlier Reconstruct. Note,
however, that line ending differences might still be a problem.