Changed default "simplified" value from "true" to "false".
A comparison of traces from Reconstruct.exe and Reconstruct.jar shows
that it is most likely appropriate to mark new traces made in the Java
version as NOT simplified. This allows these traces to be more easily
simplified in Reconstruct.exe as desired. If, however, they're already
marked as "simplified" by the Java version, then Reconstruct.exe won't
do any further simplification.