Save and load the status of "is_bezier" contour flag.
The current code supports drawing of Bezier curves in a simplified
manner. However, that information has not been stored in the XML files
supported by Reconstruct.exe. This commit stores a "type" attribute:

  <Contour ... type="bezier" ... />

The "type" attribute is only added when the contour is a bezier curve.
When no "type" is specified, it's assumed to be the standard list of
points that Reconstruct.exe has always supported. This mechanism has
allowed the storage and redrawing of bezier curves in the Java version.

This new attribute field was tested with Reconstruct.exe, and it appears
to be preserved as long as the contour itself isn't edited. When the
contour is edited, the generated XML will lose the "type" attribute.

It's not clear what's the best way to store the "handle" information
associated with each Bezier point. One method would store the handles as
points in the current points array. When opened by a non-bezier version
of reconstruct, these handles would be viewed as regular points in the
polygon. These might not be too far from realistic in some cases, but
they could be quite far off in other cases. An alternative method would
be to store the handles in another separate attribute (maybe something
like "handles"). This might be the better approach since it appears that
Reconstruct.exe will just ignore this unknown attribute. This will lose
the handles, but the points will still have the same locations (without
the handles being interpreted as regular points).