Java: Reverse All, Signed Areas, Contour(Element) constructor
It appears that Reconstruct.exe expects to find trace points in the XML
files in reverse order. This commit provides a "Reverse All" menu item
in the Trace menu to perform this reversal (typically prior to saving).

In order to check the reversal, there's also a new "Debug" option to
dump all the areas (Program/Debug/Dump Areas). This will print out all
of the areas for all of the contours for all of the sections. It's a
big help when trying to figure out the need for reversal.

A new ContourClass constructor was added to build the Contour from an
XML Element. This had been done at the Section level previously, but
since the XML is looking like it will remain the standard, the new
constructor was added to the ContourClass.

Note that Reconstruct.exe has had some trouble reading some of the
XML files generated by this version. It will show the first section,
and the scroll wheel will also show the second section, but then it
can't be moved from the second section. In all cases, showing the
section list shows only one section. No debug messages are printed
in the console either, and the XML files look OK.