Fixed drawing of Bezier curves while tracing.
The "init_bezier" function should be called whenever a point is added
and also whenever a trace is closed since these events will change the
shape of the curve.

There are still two minor problems with this version:

   1. The starting point isn't drawn until a second point is added.
   2. The starting handle isn't drawn properly until closed.

These are relatively minor problems. It might be more important to work
on editing capabilities first.