Added code to save the handles to the XML file and read back.
This version writes the handle data into a "handles" attribute of the
contour. The data is stored in the same format as the "points", but
there are 4 values per line rather than 2. This makes the number of
handles match the number of points.

At this point, editing only works for newly added points since the XML
is not updated after edits of existing points.