Added ability to save edits to points.
Previous versions allowed moving contour points, but those edits were
not saved to the XML files. This version saves those changes to the
points.

Note that points that haven't changed may be reformatted with this
version because it saves all values in full precision regardless of
whether the values have changed from the original. This should be fixed
to minimize the changes to the XML file that trigger GIT stores.