Improved the mode switching logic.
This version has two major mode controls:

   Move/Modify   and   Draw/Edit

The "Move" mode is really just zooming and panning. No data is changed
in that mode. That is the default mode and it's indicated with the large
four pointed cursor (the "Move" cursor). Zooming happens with the scroll
wheel, and paging happens with the shifted scroll wheel.

The "Modify" mode is entered with the right mouse button. There are two
ways to modify:

   Drawing new traces (Draw) shows crosshair cursor
   Editing existing traces (Edit) shows normal pointer cursor

These two modes are toggled with the shifted right mouse button. The
unshifted right mouse button still toggles between the current modify
mode and the move mode.

It's anticipated that the Edit mode might have additional hot keys for
copying (duplicating) points and deleting points. These will likely be
"c" and "x" respectively.