Working on getting the handles to work properly.
The current version generates 2 sets of handles:

   The handles stored in the "handle_points" array.
   The handles used to draw the contour.

The second set is generated dynamically during the drawing by the
"default_curve" function. This version has code to try to use the
handle_points array for drawing, but it's disabled with "false"
because it doesn't draw properly yet. At some point the values in
the handles become extremely large. It's not clear what is causing
this problem.