Previous version unintentionally modified handle_points via "hh" array
The "draw" function of ContourClass created a reference to the member
data "handle_points" via the assignment:

  hh = handle_points.get(j);

Since this wasn't a copy of the original array, modifications to "hh"
done by the translation changed the "handle_points" themselves. These
changes turned relatively small values (around 1.0) into screen
coordinate values (in the hundreds).

This modification created an additional "hht" array to hold the
transformed values needed for drawing.

Note that a number of debugging prints were also added in this commit,
but they're commented out since the problem has been found.