This version uses the "closed" flag to draw from the stored handles.
When the user is drawing a new object, it makes sense to compute the
handles automatically and to adjust all of them based on each newly
entered point. This makes sense since there is no (current) mechanism
for moving "handles" while drawing.

However, once the user has closed (or otherwise completed) an object,
then the handles can be edited, so the drawing should come from those
handles.

This code does that, except that there's an indexing problem with the
handles such that the wrong handles are connected to the wrong points.
This can be sorted out by experimenting to see which handles go with
which points. Using that technique, the curves can be controlled.

Of course, the handle indexing needs to be fixed!!