Fixed handle lines. This version works well. Allows angled joints.
Fixed the drawing of the handle lines to match the proper points. The
previous version had drawn the red lines to the wrong points.

This version supports "angled joints". In other words, there's no
enforcement of slope continuity at points. So angled corners can be
made. This might remain as an option, but the default should force slope
continuity.

This version also doesn't drag handles with their points. That's also
something that should be changed in future versions.