Added support for Bezier triplets to implement coordinated movement.
When a contour point is moved, it's most natural to move the Bezier
handles along with it. Otherwise they have to be moved separately.

Similarly, when one handle is rotated, it's often desired that the
other handle rotate opposite (preserving a straight line between the
handles passing through the contour point).

These kinds of operations require the ability to find not only the
selected points, but the points associated with them. This commit
adds that functionality through the "find_bezier_triplet" functions.
These functions take a point that is expected to be among either the
contour points or the handle points. It then returns a triplet array
containing the three points [h0, p, h1]. The indexing required to do
this is currently very clunky due to the array storage layout of the
points and handles. This should be improved in the future.