Added an option to export or not export handles to XML files.
Some of the neuropil tools programs do not accept XML files with a
"handles" attribute. It would be preferable if they could simply ignore
any fields (like "handles") that they don't recognize, but that's not an
option at this point in time. For that reason, Reconstruct Java now has
an option to not export them in the XML.