Added some nice features that were in the fix_reversing branch.
There's a problem when either loading or saving that the direction of
the points is reversed. This is due to Reconstruct itself storing the
points in reversed order. The addition of the Bezier handles complicates
the matter, and the Bezier handles are not being properly reversed to
match the points.

One consequence of this is that every time a file is saved and loaded
the direction of the points will be reversed. If there are no Bezier
handles, then that's all that happens.

However, if any of the traces were drawn with Bezier handles, then they
will appear scrambled on every other save/load cycle. It's important to
"time" this properly when adding new traces to an existing series. In
particular, the traces should look "scrambled" when adding new traces.
This will make them all match when the file is saved and reloaded. If
traces are made after loading a file that looks "unscrambled" then the
new traces and the old traces will be "out of phase" and either one or
the other will look scrambled (on alternate save/load cycles).

Another branch was created to try to address this problem, but quickly
became too complicated due to the various places (and mechanisms) used
for loading and saving traces.