Added a check on not button 3 (right click) for mouse pressed.
The Reconstruct class overloads both mouseClicked and mousePressed.

A "click" is a "higher level" event and includes both a "press" and a
"release". It was used to detect the mode change of the right button
(known as "BUTTON3" in Java). However, the "mousePressed" method was
also being called for the right button mouse "mouseClicked" event. This
change (in the "mousePressed" method) ignores right mouse presses so they
can be handled by the "mouseClicked" method.

This may have accounted for the sometimes non-responsive right clicking
behavior noted during the Texas trip.