Set up an interface to run AlignSWiFT from Reconstruct.java.
This version builds a menu from the commands available in the AlignSWiFT
class (found via get_name method). It then passes those names back to
the AlignSWiFT class when selected. This provides a very simple and very
well isolated interface between the two. This will not likely be the
final interface, but it provides a lot of flexibility at this point.