Changed paint_frame to match current Reconstruct.java behavior.
The previous version (now commented) scaled all the images to match each
other in size (at least in one dimension).

This change uses the newer Reconstruct.java code to allow images to be
of different sizes. Note that this typically centers the lower left
corner of the image (regardless of its size).