Added jiv.jar to repository, updated makefile to save it's SHA1 sum.
Note that jar files created at different times will have different SHA1
sums. This is due to a date stamp found in the .jar file. One way around
this is to force files to have a given fixed date which will then be put
in the .jar file keeping them the same. This was done for CellBlender,
and part of CellBlender's makefile was added here to show how it's done.

That wasn't implemented here because it requires making a MANIFEST.MF
file which is a bit cumbersome at this early stage of development.