Began implementing the actual qiv interface.
This version began to implement the qiv commands that were used in the
Austin Texas demonstrations of SWiFT-IR. This involved some code to
detect the wildcard file names used. However, it appears that the Linux
command line already expands wild cards into file names. It's not clear
whether this is true of Windows yet.

However, given that the Linux version won't need the wildcard handling,
that portion of the code will be put on hold for now.