Implemented a wild card file name expansion (may be needed in Windows)
It turns out that the Linux shell expands unquoted file name wild cards
into multiple command line options to the actual program.

So running:

  java -jar jiv.jar simple*.gif

will pass in simple1.gif simple2.gif and simple3.gif as arguments.

However, if the argument is quoted like this:

  java -jar jiv.jar "simple*.jpg"

then only the string "simple*.jpg" will be an argument.

Since it's not clear how other operating systems (or even other shells)
will handle wild cards, it seems wise to do internal expansion on any
file names that are passed in. This version demonstrates that. So if
there are 3 simple#.jpg files and 3 simple#.gif files, this command:

  java -jar jiv.jar -D "simple*.jpg" simple*.gif

will match all 6 files.