JIV: Connected the command line arguments to the actual image loading.
This version will load the images specified on the command line using
the standard wild card expansion by the shell and further expanding any
"*" or "?" wild cards that are not processed by the shell. This should
hopefully work in most operating systems.

Note that this version attempts to load all of the images at once. That
could be easily modified by a boolean expression passed to the jiv_frame
constructor in the main program. If that parameter were "i<3", then only
the first 3 images would actually be loaded at startup. Then other
images could be loaded on demand as currently done in Reconstruct.java.
But to keep things simple, this version just tries to load them all at
once.