Move the refresh button from the file menu to the menu bar itself.
This appears to work, but the menu bar item isn't displayed the same as
the other drop down menus. It may be that the Swing components are not
intended to have actual menu items placed in the menu bar. In this case,
the style of the menu item is slightly different and the menu item seems
to force the "help" menu to the far right. This kind of single click
functionality is better provided via a tool bar, but that's more work
than needed for now.

If having the refresh in the menu bar is a problem, simply reverse the
changes in this commit.