Added a line to print the current frame title to the console.
The "set_title" function builds a title string from the current file
and frame number. The "set_title" function then calls the the parent
frame's "setTitle" function. This works on some window managers, but
not on others. As a quick backup, this change will display the current
frame number and file name on the console.