Set frame index to first new file added to the stack.
The previous version always set the frame index to the last file that
had been added. This was true whether there were any existing images or
not.

This change sets the frame index to the first new image added. This is
true whether there were any existing images or not.