Added preliminary "glitch" detection code that looks for sharp angles
This version adds a "Show Glitch" menu option to the "Extras" menu. When
enabled, glitches are detected and reported when showing a section. This
happens at draw time which is appropriate for making corrections.

The current "glitch" detection looks for consecutive points with a
normalized angle less than 0.3 radians (about 17 degrees). This angle
was arbitrarily selected and may be added as an option in the future.

Also note that this version only compares 3 consecutive points to look
for a sharp angle. This may miss certain types of glitches.