Java: Adding on a "Delete Point" mode. Limit Select to Current Section.
While working to add a "Delete Point" mode, it was discovered that the
current point selection mechanism was searching for match points across
all sections in the function SeriesClass.find_closest(). This is not
what is generally desired, but it will tend to find the desired point if
the stack of contours is somewhat sparse (lower chance of finding a
closer point in another section). This commit changed the functionality
of the SeriesClass.find_closest() to find the closest for the current
section only. It also retained the previous functionality in the
function SeriesClass.find_closest_all() as well as another new overload
to SeriesClass.find_closest() that accepts a second section_index
parameter.