Java: Same changes for find_bezier_triplet as find_closest
As with find_closest, the find_bezier_triplet was operating over the
entire series rather than on the current section. As with find_closest,
this commit changes the default to only search on the current section.
It also adds a function to search any given section and a search all
sections function (which had been the default).