Java: Made it easier to start empty projects with no images.
To create a project with 5 empty sections:

  Series / New  (give it a new name ending with ".ser")
  Program / Debug / New Empty Sections

It will complain that it cannot open the file named:

    "File_that_does_not_exist.jpg"

Click "OK" to dismiss the dialog box.

The project will now have 5 empty sections with no images.
This can be useful for testing various drawing capabilities.

Note that for adding and deleting points, Bezier drawing must be off:

  Extras / Mode / Bezier Drawing