Added the Add/Delete "mode" feature toggled with 'a' and 'x'
This version uses the 'a' and 'x' keys (also 'i' and 'd') to toggle the
mode into and out of the add and delete modes respectively. Right now
the 'a' key (add) and 'i' key (insert) perform the same function. The
same is true of the 'x' and 'd' (delete) keys. At some point the dual
mapping may be discontinued to use the keys for other purposes. It's
likely that 'a' and 'x' will remain.