Added zoom limits via scroll wheel and attempted to widen '+' cursor
Without zoom limits, it's possible for the scaling values to exceed
certain numerical limits related to the Java display. This change limits
the scroll wheel to +/- 76. This limits the actual scaling to a maximum
of about 4.56e8 and a minimum of about 2.18e-9. These should be fine for
viewing most images.