Reconstruct Java: Modified makefile to create a command line reconstruct
It appears to be possible to add the needed text to the front of a .jar
file ("#!/usr/bin/java -jar") to make a jar file executable from the
command line on unix-like systems. The discription is available here:

   https://stackoverflow.com/questions/1667830

The instructions from that URL were implemented in "makefile" to generate
an executable "reconstruct" that runs from the command line (assuming
that "/usr/bin/java" is a working Java installation). This may not be
particularly useful for Reconstruct (which isn't often run as a command
line program), but it may be useful for other Java based utilities.